requirements.txt
matplotlib==3.10.1
numpy==2.1.3
pandas==2.2.3
scanpy==1.11.2
scikit-learn==1.6.1
scikit-image==0.25.2
scipy==1.15.2
scTAPE==1.1.2
torch==2.6.0
tqdm==4.67.1
igraph==0.11.8
